Aramco_AIS_Repo_updated_with_src/Aramco_AIS_Repo/P1_Compressor_Anomaly/requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.2
matplotlib>=3.7
fastapi>=0.103
uvicorn>=0.23
pydantic>=2.1
joblib>=1.3
torch>=2.0
prometheus-client>=0.16
mlflow>=2.3
pyyaml>=5.4